Add files via upload
This project builds a voice-controlled car that can be controlled by voice which react in accordance with the corresponding voice command. However, noise and distance handling require future development. Simple voice commands like left, right, forward, back, stop is used to run the car. These commands are given to Bluetooth module via an android application. The Bluetooth module and control unit are combined to store and test the voice commands.  When  any  instruction  for  the  automobile  (car)  is  identified,  a command message is sent to Arduino UNO, the Microcontroller of the car by the Bluetooth  device.  This  command  is  analyzed  by  the  microcontroller  and followed up. In the vehicle. This work has been limited to short-range (100m range) and is linked to the car over long distance via long-range modules.